# backend/docs/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "New API",
        "version": "1.0.0",
        "description": "This is a swagger ui for out trennkodus API"
    },
    "paths": {
        "/exercises": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercise"
                                },
                                "examples": {
                                    "List of exercises": {
                                        "value": {
                                            "name": "some text",
                                            "set": 0,
                                            "repesition": 60,
                                            "id": 21
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Get all Exercise names"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewExercise"
                            },
                            "examples": {
                                "NewExercise": {
                                    "value": {
                                        "name": "some text",
                                        "set": 56,
                                        "repesition": 24
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Exercises"
                ],
                "responses": {
                    "201": {
                        "headers": {
                            "location": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Location of created object"
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercise"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Add new exercise"
            }
        },
        "/exercises/{id}": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "examples": {
                                    "Full details of a given exercise": {
                                        "value": {
                                            "id": 1,
                                            "name": "push-up",
                                            "repetition": 10,
                                            "set": 3
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Returns full details of a given id exercise"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "Exercise /w+ was not found": {
                                        "value": {
                                            "error": "Exercise \\w+ was not found in exercise array."
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Exercise was not found in exercises array"
                    }
                },
                "operationId": "id",
                "summary": "Get full details of a exercise",
                "description": "Returns full details of a given id exercise"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewExercise"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Exercises"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercise"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Modify existing exercise"
            },
            "delete": {
                "tags": [
                    "Exercises"
                ],
                "responses": {
                    "204": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Exercise deleted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "IdError": {
                                        "value": {
                                            "error": "ID must be a whole number: esimene"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Incorrect Id"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "NotFound": {
                                        "value": {
                                            "error": "Exercise not found"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Exercise not found"
                    }
                },
                "summary": "Remove a exercise"
            },
            "parameters": [
                {
                    "name": "id",
                    "description": "Id of the exercise to be modified",
                    "schema": {
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/trainingPrograms": {
            "get": {
                "tags": [
                    "Trainingprogram"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrainingPrograms"
                                },
                                "examples": {
                                    "List of training programs": {
                                        "value": {
                                            "Exercises": {
                                                "id": 38,
                                                "name": "some text",
                                                "repetisions": 96,
                                                "Sets": 76
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Get a list of Training Programs"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewTrainingProgram"
                            },
                            "examples": {
                                "NewTrainingProgram": {
                                    "value": {
                                        "day": "some text",
                                        "Exercises": [
                                            {
                                                "id": 1,
                                                "name": "some text",
                                                "repetisions": 54,
                                                "Sets": 93
                                            },
                                            {
                                                "id": 2,
                                                "name": "some text",
                                                "repetisions": 54,
                                                "Sets": 93
                                            },
                                            {
                                                "id": 3,
                                                "name": "some text",
                                                "repetisions": 54,
                                                "Sets": 93
                                            },
                                            {
                                                "id": 4,
                                                "name": "some text",
                                                "repetisions": 54,
                                                "Sets": 93
                                            },
                                            {
                                                "id": 5,
                                                "name": "some text",
                                                "repetisions": 54,
                                                "Sets": 93
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Trainingprogram"
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrainingPrograms"
                                }
                            }
                        },
                        "description": "Success"
                    }
                },
                "summary": "Add new training program"
            }
        },
        "/trainingPrograms/{Id}": {
            "get": {
                "tags": [
                    "Trainingprogram"
                ],
                "parameters": [
                    {
                        "name": "Id",
                        "description": "Id of the training program to be modified",
                        "schema": {
                            "type": "integer"
                        },
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "examples": {
                                    "Full details of a given training program": {
                                        "value": "    {\n        \"id\": 1,\n        \"day\": \"Chest day\",\n        \"exercises\": [\n            {\n                \"id\": 1,\n                \"name\": \"push-up\",\n                \"repetitions\": 10,\n                \"sets\": 3\n            },\n            {\n                \"id\": 2,\n                \"name\": \"Diamond push-up\",\n                \"repetitions\": 10,\n                \"sets\": 3\n            },\n            {\n                \"id\": 3,\n                \"name\": \"squat\",\n                \"repetitions\": 10,\n                \"sets\": 3\n            },\n            {\n                \"id\": 4,\n                \"name\": \"pistol squat\",\n                \"repetitions\": 10,\n                \"sets\": 3\n            },\n            {\n                \"id\": 5,\n                \"name\": \"lat pull down\",\n                \"repetitions\": 10,\n                \"sets\": 3\n            }\n        ]\n    }"
                                    }
                                }
                            }
                        },
                        "description": "Returns full details of a given id training program"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                },
                                "examples": {
                                    "Training program was not found": {
                                        "value": {
                                            "error": "Training program was not found in training program array"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Training program was not found in training programs array"
                    }
                },
                "operationId": "Id",
                "summary": "Get full details of a training program",
                "description": "Returns full details of a given id training program"
            },
            "parameters": [
                {
                    "name": "Id",
                    "description": "Id of the training program to be modified",
                    "schema": {
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        }
    },
    "components": {
        "schemas": {
            "error": {
                "title": "Root Type for error",
                "description": "",
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "example": {
                    "error": "Error message."
                }
            },
            "Exercise": {
                "title": "Root Type for new-exercise",
                "description": "Creates a new exercise",
                "required": [
                    "id",
                    "name",
                    "repetisions",
                    "Sets"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "format": "int32",
                        "description": "",
                        "type": "integer"
                    },
                    "name": {
                        "description": "",
                        "type": "string"
                    },
                    "repetisions": {
                        "format": "int32",
                        "description": "",
                        "type": "integer"
                    },
                    "Sets": {
                        "format": "int32",
                        "description": "",
                        "type": "integer"
                    }
                },
                "example": {
                    "id": 67,
                    "name": "some text",
                    "repetitions": 43,
                    "Sets": 9
                }
            },
            "NewExercise": {
                "description": "Exercise to be saved",
                "required": [
                    "name",
                    "repetitions",
                    "sets"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "description": "Must contain atleast one non-whitespace character",
                        "type": "string"
                    },
                    "sets": {
                        "format": "int32",
                        "description": "",
                        "type": "integer"
                    },
                    "repetitions": {
                        "format": "int32",
                        "description": "",
                        "type": "integer"
                    }
                }
            },
            "TrainingProgram": {
                "description": "",
                "required": [
                    "Id"
                ],
                "type": "object",
                "properties": {
                    "Exercises": {
                        "$ref": "#/components/schemas/Exercise",
                        "description": ""
                    },
                    "day": {
                        "description": "",
                        "type": "string"
                    },
                    "Id": {
                        "description": "",
                        "type": "integer"
                    }
                },
                "example": {
                    "Id": 1,
                    "day": "some text",
                    "Exercises": [
                        {
                            "id": 1,
                            "name": "some text",
                            "repetisions": 54,
                            "Sets": 93
                        },
                        {
                            "id": 2,
                            "name": "some text",
                            "repetisions": 54,
                            "Sets": 93
                        },
                        {
                            "id": 3,
                            "name": "some text",
                            "repetisions": 54,
                            "Sets": 93
                        },
                        {
                            "id": 4,
                            "name": "some text",
                            "repetisions": 54,
                            "Sets": 93
                        },
                        {
                            "id": 5,
                            "name": "some text",
                            "repetisions": 54,
                            "Sets": 93
                        }
                    ]
                }
            },
            "NewTrainingProgram": {
                "description": "",
                "type": "object",
                "properties": {
                    "day": {
                        "description": "",
                        "type": "string"
                    },
                    "exercises": {
                        "$ref": "#/components/schemas/Exercise",
                        "description": ""
                    }
                }
            }
        }
    }
}